==> Code/1 - Hydrate Tweets.ipynb <==
# COVID-19 Twitter NLP
### *Ayush Noori, Omar Wahby, Ethan Fang*

# !conda info --envs
# !where python

First, load all requisite libraries.

# base libraries
import numpy as np
import pandas as pd
import configparser
import os

# download files
from datetime import datetime
import wget

# Twitter libraries
import jsonlines, json
from twarc import Twarc
import csv

# base directory
base_dir = os.getcwd()
print("Base Directory: " + base_dir)

## Read API Keys
Read API keys for Twitter authentication.

config = configparser.ConfigParser()
config.read("config.ini")
api_keys = config["twitter"]

# read keys from config file
consumer_key = api_keys["consumer_key"]
consumer_secret = api_keys["consumer_secret"]
access_token = api_keys["access_token"]
access_token_secret = api_keys["access_token_secret"]

# create Twarc object
t = Twarc(consumer_key, consumer_secret, access_token, access_token_secret)

# Download Files
First, indicate the date range of the files you would like to hydrate. Rememeber that `start_date` cannot be before `2020-01-22`.

import pandas as pd
start_date = '2021-07-01'
end_date = '2021-07-01'

datelist = pd.date_range(start_date, end_date).tolist()
number_files = 0
for date in datelist:
  for h in ['00','01', '02','03','04','05','06','07','08','09','10','11','12','13','14','15','16','17','18','19','20','21','22','23']:
    number_files = number_files + 1

print("You have requested to download a total of {} files.".format(number_files))

summary_dir = os.path.join(base_dir, "..//Data//Summary Details")
os.chdir(summary_dir)

datelist = pd.date_range(start_date, end_date).tolist()
files_list = []
for date in datelist:
  day = date.strftime("%Y_%m_%d")
  p1 = "https://raw.githubusercontent.com/lopezbec/COVID19_Tweets_Dataset/master/Summary_Details/"
  m = day[0:7]
  for h in ['00','01', '02','03','04','05','06','07','08','09','10','11','12','13','14','15','16','17','18','19','20','21','22','23']:
    file = p1 + m + "/" + day + "_" + str(h) + "_Summary_Details.csv"
    wget.download(file)

# Filter Tweet IDs
Filter Tweets by specific parameters.

summary_details = pd.DataFrame() 
directory =  os.getcwd()
for root, dirs, files in os.walk(directory):
  for file in files:
    if file.endswith(".csv"):
      data = pd.read_csv(file) 
      frames = [summary_details, data]
      summary_details = pd.concat(frames)
summary_details

Here, we filter to include only tweets that are in English are NOT retweets (i.e., are original tweets). We could also filter for only tweets with geolocation coordinates.

summary_details_filter = summary_details
summary_details_filter = summary_details_filter[summary_details_filter['Language']=='en']
# summary_details_filter = summary_details_filter[summary_details_filter['Geolocation_coordinate']=='YES']
summary_details_filter = summary_details_filter[summary_details_filter['RT']=='NO']
summary_details_filter


data_dir = os.path.join(base_dir, "..//Data//")
os.chdir(data_dir)
final_tweet_ids_filename = "Filtered Tweet IDs.txt"

# store final IDs
with open(final_tweet_ids_filename, "w+") as f:
    for id in summary_details_filter['Tweet_ID']:
        f.write('%s\n' % id)

# Hydrate
Now, hydrate the tweets using the selected Tweet IDs.

final_tweet_ids_filename = "Filtered Tweet IDs.txt"
output_filename = "Hydrated Tweets.csv"

The time required to run the below chunk will depend on the number of tweets one would like to hydrate.

# stores hydrated tweets here as jsonl objects
# contains one json object per line
output_json_filename = output_filename[:output_filename.index(".")] + ".txt"
ids = []
with open(final_tweet_ids_filename, "r") as ids_file:
    ids = ids_file.read().split()
hydrated_tweets = []
ids_to_hydrate = set(ids)

# looks at the output file for already hydrated tweets
if os.path.isfile(output_json_filename):
    with jsonlines.open(output_json_filename, "r") as reader:
        for i in reader.iter(type=dict, skip_invalid=True):
            # these tweets have already been hydrated. So remove them from ids_to_hydrate
            hydrated_tweets.append(i)
            ids_to_hydrate.remove(i["id_str"])
print("Total IDs: " + str(len(ids)) + ", IDs to hydrate: " + str(len(ids_to_hydrate)))
print("Hydrated: " + str(len(hydrated_tweets)))

count = len(hydrated_tweets)
start_index = count # the index from where tweets haven't been saved to the output_json_file
# stores hydrated tweets to output_json_file every num_save iterations.
num_save  = 1000

# now, use twarc and start hydrating
for tweet in t.hydrate(ids_to_hydrate):
    hydrated_tweets.append(tweet)
    count += 1
    # if num_save iterations have passed
    if (count % num_save) == 0:
        # open the output file
        # NOTE: even if the code stops during IO, only tweets from the current iteration are lost
        # older tweets are preserved as the file is written in append mode
        with jsonlines.open(output_json_filename, "a") as writer:
            ## print("Started IO")
            # now write the tweets from start_index
            # the other tweets don't have to be written as they were already written in a previous iteration or run.
            for hydrated_tweet in hydrated_tweets[start_index:]:
                writer.write(hydrated_tweet)
            ## print("Finished IO")
        print("Saved " + str(count) + " hydrated tweets.")
        # now, since everything has been written, reset start_index
        start_index = count
# there might be tweets unwritten in the last iteration if the count is not a multiple of num_tweets
# in that case, just write out the remainder of tweets
if count != start_index:
    print("Here with start_index", start_index)
    with jsonlines.open(output_json_filename, "a") as writer:
        for hydrated_tweet in hydrated_tweets[start_index:]:
           writer.write(hydrated_tweet)

## Convert `JSONL` to `CSV`

output_json_filename = output_filename[:output_filename.index(".")] + ".txt"

# these are the column name that are selected to be stored in the csv
keyset = ["created_at", "id", "id_str", "full_text", "source", "truncated", "in_reply_to_status_id",
          "in_reply_to_status_id_str", "in_reply_to_user_id", "in_reply_to_user_id_str", 
          "in_reply_to_screen_name", "user", "coordinates", "place", "quoted_status_id",
          "quoted_status_id_str", "is_quote_status", "quoted_status", "retweeted_status", 
          "quote_count", "reply_count", "retweet_count", "favorite_count", "entities", 
          "extended_entities", "favorited", "retweeted", "possibly_sensitive", "filter_level", 
          "lang", "matching_rules", "current_user_retweet", "scopes", "withheld_copyright", 
          "withheld_in_countries", "withheld_scope", "geo", "contributors", "display_text_range",
          "quoted_status_permalink"]
hydrated_tweets = []

# reads the current tweets
with jsonlines.open(output_json_filename, "r") as reader:
    for i in reader.iter(type = dict, skip_invalid = True):
        hydrated_tweets.append(i)

# convert nested dictionary to JSON format for R
tweet_n = 0
for tweet in hydrated_tweets:
    tweet['user'] = json.dumps(tweet['user'])
    tweet['entities'] = json.dumps(tweet['entities'])
    # tweet['quoted_status_permalink'] = json.dumps(tweet['quoted_status_permalink'])
    tweet['place'] = json.dumps(tweet['place'])
    tweet['coordinates'] = json.dumps(tweet['coordinates'])
    tweet_n = tweet_n + 1
    if (tweet_n % 50000) == 0:
        print("Parsed {} tweets.".format(tweet_n))

# write tweets out
with open(output_filename, "w+", encoding = "utf-8") as output_file:
    d = csv.DictWriter(output_file, keyset)
    d.writeheader()
    d.writerows(hydrated_tweets)
