==> Code/4 - STM Visualization.Rmd <==
---
title: STM Visualization
subtitle: Ayush Noori
output:
  prettydoc::html_pretty:
    theme: cayman
    highlight: github
    toc: yes
editor_options: 
  chunk_output_type: inline
---

# Dependencies

Load requisite packages and define directories. Note that this script loads my personal utilities package, `ayushnoori/brainstorm`, which is available via GitHub.

```{r load-packages, message=FALSE, warning=FALSE}

# data manipulation
library(data.table)
library(purrr)
library(magrittr)
library(stringr)

# structural topic modeling
library(stm)

# data visualization
library(ggplot2)

# source utilities script
library(brainstorm)

```

Directories are relative to the R project path.

```{r define-directores}

bdir = file.path("Results", "2 - Parse Tweets")
ddir = file.path("Results", "3 - Semantic Analysis")
rdir = file.path("Results", "4 - STM Visualization")

```

# Read Data

```{r read-data}

twt = readRDS(file.path(bdir, "Final Parsed Tweets.RDS"))
processed = readRDS(file.path(ddir, "Processed Text.RDS"))
out = readRDS(file.path(ddir, "Prepared Documents.RDS"))
stm_fit = readRDS(file.path(ddir, "Structural Topic Model K56.RDS"))

```

Visualize interesting topics and find representative documents.

```{r representative-documents}

sel_topics = c(3, 5, 9, 11, 14, 21, 23, 32, 34, 37, 41, 55)

rep_docs = function(topic_n) {
  
  thoughts_n = findThoughts(stm_fit, texts = out$meta$CleanText, n = 100, topics = topic_n)
  thoughts_df = data.frame(topic_n, thoughts_n$index, thoughts_n$docs)
  colnames(thoughts_df) = c("Topic", "Index", "Text")
  thoughts_df = cbind(thoughts_df, out$meta[thoughts_df$Index, ])
  
}

```

Map function over interesting topics to get representative documents.

```{r map-representative}

rep_list = map_dfr(sel_topics, rep_docs)
setDT(rep_list)

rep_list %>%
  .[, ID := bit64::as.character.integer64(ID)] %>%
  .[, c("FullText", "CleanText") := NULL] %>%
  .[, TweetURL := paste0("https://twitter.com/", UserScreenName, "/status/", ID)] %>%
  setcolorder("TweetURL")

saveRDS(rep_list, file.path(rdir, "Representative Tweets.RDS"))
write_excel(rep_list, file.path(rdir, "Representative Tweets.xlsx"),
            sheet = "Representative Tweets", overwrite = F)

```

Plot topics summary.

```{r plot-topics}

pdf(file.path(rdir, "Top Topics.pdf"), width = 16, height = 16)
plot(stm_fit, type = "summary", xlim = c(0, 0.1))
dev.off()

# stm_corr = topicCorr(stm_fit)
# plot(stm_corr)

```

Check correlation with followers and friends.

```{r check-correlation}

prep = estimateEffect(c(3, 5, 9, 14, 34) ~ UserFollowers + UserFriends, stm_fit, meta = out$meta, uncertainty = "None")
summary(prep)

prep_followers = estimateEffect(c(3, 5, 9, 14, 34) ~ s(UserFollowers), stm_fit, meta = out$meta, uncertainty = "None")

prep_friends = estimateEffect(c(3, 5, 9, 14, 34) ~ s(UserFriends), stm_fit, meta = out$meta, uncertainty = "None")

#labeltype = "prob", model = stm_fit, 
pdf(file.path(rdir, "Followers Relationship.pdf"), width = 8, height = 8)
plot(prep_followers, "UserFollowers", method = "continuous", topics = c(3, 5, 9, 14, 34), xlab = "Followers", ci.level = NA, ylim = c(-0.07, 0.1), xlim = c(1e07/2, 6e07), linecol = c("#70A2FF", "#E07A5F", "#3D405B", "#67A286", "#EEBC6D"))
dev.off()

pdf(file.path(rdir, "Friends Relationship.pdf"), width = 8, height = 8)
plot(prep_friends, "UserFriends", method = "continuous", topics = c(3, 5, 9, 14, 34), xlab = "Friends", ci.level = NA, ylim = c(-0.2, 0.2), xlim = c(100000, 1900000), linecol = c("#70A2FF", "#E07A5F", "#3D405B", "#67A286", "#EEBC6D"))
dev.off()

```
